Fixing of ruff issues + update on the intercompatibility for python version in the CI pipeline

+++ alertwest_scraping/send_annotation_api.py
@@ -57,7 +57,7 @@ def import_sequence_via_annotation_api(
     sequence,
     cam_id: str,
     cam_name: str,
-    azimuth: Optional[int],
+    azimuth: Optional[float],
     lat: float,
     lon: float,
     alert_api_id: int,
